refactored more commands
