+++ frontend/src/ts/commandline/lists/random-theme.ts
@@ -1,0 +1,67 @@
+import * as UpdateConfig from "../../config";
+import { Auth } from "../../firebase";
+import * as Notifications from "../../elements/notifications";
+
+const commands: MonkeyTypes.CommandsGroup = {
+  title: "Random theme...",
+  configKey: "randomTheme",
+  list: [
+    {
+      id: "setRandomOff",
+      display: "off",
+      configValue: "off",
+      exec: (): void => {
+        UpdateConfig.setRandomTheme("off");
+      },
+    },
+    {
+      id: "setRandomOn",
+      display: "on",
+      configValue: "on",
+      exec: (): void => {
+        UpdateConfig.setRandomTheme("on");
+      },
+    },
+    {
+      id: "setRandomFav",
+      display: "fav",
+      configValue: "fav",
+      exec: (): void => {
+        UpdateConfig.setRandomTheme("fav");
+      },
+    },
+    {
+      id: "setRandomLight",
+      display: "light",
+      configValue: "light",
+      exec: (): void => {
+        UpdateConfig.setRandomTheme("light");
+      },
+    },
+    {
+      id: "setRandomDark",
+      display: "dark",
+      configValue: "dark",
+      exec: (): void => {
+        UpdateConfig.setRandomTheme("dark");
+      },
+    },
+    {
+      id: "setRandomCustom",
+      display: "custom",
+      configValue: "custom",
+      exec: (): void => {
+        if (Auth.currentUser === null) {
+          Notifications.add(
+            "Multiple custom themes are available to logged in users only",
+            0
+          );
+          return;
+        }
+        UpdateConfig.setRandomTheme("custom");
+      },
+    },
+  ],
+};
+
+export default commands;
