+++ frontend/src/ts/elements/commandline-lists.ts
@@ -50,99 +50,6 @@ const commandsKeymapMode: MonkeyTypes.CommandsGroup = {
   ],
 };
 
-const commandsRandomTheme: MonkeyTypes.CommandsGroup = {
-  title: "Random theme...",
-  configKey: "randomTheme",
-  list: [
-    {
-      id: "setRandomOff",
-      display: "off",
-      configValue: "off",
-      exec: (): void => {
-        UpdateConfig.setRandomTheme("off");
-      },
-    },
-    {
-      id: "setRandomOn",
-      display: "on",
-      configValue: "on",
-      exec: (): void => {
-        UpdateConfig.setRandomTheme("on");
-      },
-    },
-    {
-      id: "setRandomFav",
-      display: "fav",
-      configValue: "fav",
-      exec: (): void => {
-        UpdateConfig.setRandomTheme("fav");
-      },
-    },
-    {
-      id: "setRandomLight",
-      display: "light",
-      configValue: "light",
-      exec: (): void => {
-        UpdateConfig.setRandomTheme("light");
-      },
-    },
-    {
-      id: "setRandomDark",
-      display: "dark",
-      configValue: "dark",
-      exec: (): void => {
-        UpdateConfig.setRandomTheme("dark");
-      },
-    },
-    {
-      id: "setRandomCustom",
-      display: "custom",
-      configValue: "custom",
-      exec: (): void => {
-        if (Auth.currentUser === null) {
-          Notifications.add(
-            "Multiple custom themes are available to logged in users only",
-            0
-          );
-          return;
-        }
-        UpdateConfig.setRandomTheme("custom");
-      },
-    },
-  ],
-};
-
-const commandsDifficulty: MonkeyTypes.CommandsGroup = {
-  title: "Difficulty...",
-  configKey: "difficulty",
-  list: [
-    {
-      id: "setDifficultyNormal",
-      display: "normal",
-      configValue: "normal",
-      exec: (): void => {
-        UpdateConfig.setDifficulty("normal");
-      },
-    },
-    {
-      id: "setDifficultyExpert",
-      display: "expert",
-      configValue: "expert",
-      exec: (): void => {
-        UpdateConfig.setDifficulty("expert");
-      },
-    },
-    {
-      id: "setDifficultyMaster",
-      display: "master",
-      configValue: "master",
-      exec: (): void => {
-        UpdateConfig.setDifficulty("master");
-      },
-    },
-  ],
-};
-
 export const commandsEnableAds: MonkeyTypes.CommandsGroup = {
   title: "Set enable ads...",
   configKey: "ads",
@@ -182,166 +89,6 @@ export const commandsEnableAds: MonkeyTypes.CommandsGroup = {
   ],
 };
 
-export const customThemeListCommands: MonkeyTypes.CommandsGroup = {
-  title: "Custom themes list...",
-  // configKey: "customThemeId",
-  list: [],
-};
-
-export function updateCustomThemeListCommands(): void {
-  if (Auth.currentUser === null) {
-    return;
-  }
-
-  customThemeListCommands.list = [];
-
-  const snapshot = DB.getSnapshot();
-
-  if (!snapshot) return;
-
-  if (DB.getSnapshot().customThemes.length === 0) {
-    return;
-  }
-  DB.getSnapshot().customThemes.forEach((theme) => {
-    customThemeListCommands.list.push({
-      id: "setCustomThemeId" + theme._id,
-      display: theme.name,
-      configValue: theme._id,
-      hover: (): void => {
-        ThemeController.preview(theme._id, true);
-      },
-      exec: (): void => {
-        // UpdateConfig.setCustomThemeId(theme._id);
-        UpdateConfig.setCustomTheme(true);
-        ThemeController.set(theme._id, true);
-      },
-    });
-  });
-  return;
-}
-
-const commandsCaretStyle: MonkeyTypes.CommandsGroup = {
-  title: "Change caret style...",
-  configKey: "caretStyle",
-  list: [
-    {
-      id: "setCaretStyleOff",
-      display: "off",
-      configValue: "off",
-      exec: (): void => {
-        UpdateConfig.setCaretStyle("off");
-      },
-    },
-    {
-      id: "setCaretStyleDefault",
-      display: "line",
-      configValue: "default",
-      exec: (): void => {
-        UpdateConfig.setCaretStyle("default");
-      },
-    },
-    {
-      id: "setCaretStyleBlock",
-      display: "block",
-      configValue: "block",
-      exec: (): void => {
-        UpdateConfig.setCaretStyle("block");
-      },
-    },
-    {
-      id: "setCaretStyleOutline",
-      display: "outline-block",
-      configValue: "outline",
-      exec: (): void => {
-        UpdateConfig.setCaretStyle("outline");
-      },
-    },
-    {
-      id: "setCaretStyleUnderline",
-      display: "underline",
-      configValue: "underline",
-      exec: (): void => {
-        UpdateConfig.setCaretStyle("underline");
-      },
-    },
-    {
-      id: "setCaretStyleCarrot",
-      display: "carrot",
-      configValue: "carrot",
-      visible: false,
-      exec: (): void => {
-        UpdateConfig.setCaretStyle("carrot");
-      },
-    },
-    {
-      id: "setCaretStyleBanana",
-      display: "banana",
-      configValue: "banana",
-      visible: false,
-      exec: (): void => {
-        UpdateConfig.setCaretStyle("banana");
-      },
-    },
-  ],
-};
-
-const commandsPaceCaretStyle: MonkeyTypes.CommandsGroup = {
-  title: "Change pace caret style...",
-  configKey: "paceCaretStyle",
-  list: [
-    {
-      id: "setPaceCaretStyleDefault",
-      display: "line",
-      configValue: "default",
-      exec: (): void => {
-        UpdateConfig.setPaceCaretStyle("default");
-      },
-    },
-    {
-      id: "setPaceCaretStyleBlock",
-      display: "block",
-      configValue: "block",
-      exec: (): void => {
-        UpdateConfig.setPaceCaretStyle("block");
-      },
-    },
-    {
-      id: "setPaceCaretStyleOutline",
-      display: "outline-block",
-      configValue: "outline",
-      exec: (): void => {
-        UpdateConfig.setPaceCaretStyle("outline");
-      },
-    },
-    {
-      id: "setPaceCaretStyleUnderline",
-      display: "underline",
-      configValue: "underline",
-      exec: (): void => {
-        UpdateConfig.setPaceCaretStyle("underline");
-      },
-    },
-    {
-      id: "setPaceCaretStyleCarrot",
-      display: "carrot",
-      configValue: "carrot",
-      visible: false,
-      exec: (): void => {
-        UpdateConfig.setPaceCaretStyle("carrot");
-      },
-    },
-    {
-      id: "setPaceCaretStyleBanana",
-      display: "banana",
-      configValue: "banana",
-      visible: false,
-      exec: (): void => {
-        UpdateConfig.setPaceCaretStyle("banana");
-      },
-    },
-  ],
-};
-
 const commandsRepeatedPace: MonkeyTypes.CommandsGroup = {
   title: "Repeated pace...",
   configKey: "repeatedPace",
@@ -360,61 +107,6 @@ const commandsRepeatedPace: MonkeyTypes.CommandsGroup = {
       configValue: true,
       exec: (): void => {
         UpdateConfig.setRepeatedPace(true);
-      },
-    },
-  ],
-};
-
-const commandsPaceCaret: MonkeyTypes.CommandsGroup = {
-  title: "Pace caret mode...",
-  configKey: "paceCaret",
-  list: [
-    {
-      id: "setPaceCaretOff",
-      display: "off",
-      configValue: "off",
-      exec: (): void => {
-        UpdateConfig.setPaceCaret("off");
-        TestLogic.restart();
-      },
-    },
-    {
-      id: "setPaceCaretPb",
-      display: "pb",
-      configValue: "pb",
-      exec: (): void => {
-        UpdateConfig.setPaceCaret("pb");
-        TestLogic.restart();
-      },
-    },
-    {
-      id: "setPaceCaretLast",
-      display: "last",
-      configValue: "last",
-      exec: (): void => {
-        UpdateConfig.setPaceCaret("last");
-        TestLogic.restart();
-      },
-    },
-    {
-      id: "setPaceCaretAverage",
-      display: "average",
-      configValue: "average",
-      exec: (): void => {
-        UpdateConfig.setPaceCaret("average");
-        TestLogic.restart();
-      },
-    },
-    {
-      id: "setPaceCaretCustom",
-      display: "custom...",
-      configValue: "custom",
-      input: true,
-      exec: (input): void => {
-        if (!input) return;
-        UpdateConfig.setPaceCaretCustomSpeed(parseInt(input));
-        UpdateConfig.setPaceCaret("custom");
-        TestLogic.restart();
       },
     },
   ],
